+++ doc/release_notes.md
@@ -2,6 +2,10 @@
 
 ## rust-oracledb 26.0.0-beta.2 (TBD)
 
+1.  Added method [Statement::bind_names()](crate::Statement::bind_names()) in
+    order to determine the list of bind variable names used by a statement.
+1.  Fixed bugs and enhanced parsing of SQL statements
+    ([issue 1](https://github.com/oracle/rust-oracledb/issues/1)).
 1.  Fixed bugs and enhanced parsing of connect strings
     ([issue 2](https://github.com/oracle/rust-oracledb/issues/2)).
 1.  Added runnable examples.